+++ controllers/admin/AdminDPDBalticsAjaxShipmentsController.php
@@ -17,6 +17,7 @@ use Invertus\dpdBaltics\DTO\ShipmentData;
 use Invertus\dpdBaltics\Factory\ShipmentDataFactory;
 use Invertus\dpdBaltics\Repository\ShipmentRepository;
 use Invertus\dpdBaltics\Service\Address\ReceiverAddressService;
+use Invertus\dpdBaltics\Service\API\LabelApiService;
 use Invertus\dpdBaltics\Service\API\ParcelShopSearchApiService;
 use Invertus\dpdBaltics\Service\API\ShipmentApiService;
 use Invertus\dpdBaltics\Service\Exception\ExceptionService;
@@ -30,12 +31,13 @@ use Invertus\dpdBalticsApi\Api\DTO\Response\ParcelShopSearchResponse;
 use Invertus\dpdBalticsApi\Api\DTO\Response\ShipmentCreationResponse;
 use Invertus\dpdBalticsApi\Exception\DPDBalticsAPIException;
 
-require_once dirname(__DIR__).'/../vendor/autoload.php';
+require_once dirname(__DIR__) . '/../vendor/autoload.php';
 
 class AdminDPDBalticsAjaxShipmentsController extends AbstractAdminController
 {
     protected $ajaxActions = ['save', 'save_and_print', 'updateAddressBlock', 'getProductPriceByID', 'print'];
     protected $ajaxPudoActions = ['searchPudoServices'];
+    const DEFAULT_LABEL_TEMPLATE_ID = 1;
 
     /**
      * Process AJAX call
@@ -43,7 +45,6 @@ class AdminDPDBalticsAjaxShipmentsController extends AbstractAdminController
     public function postProcess()
     {
         $response = ['status' => false];
-
         $action = Tools::getValue('action');
         $idOrder = Tools::getValue('id_order');
         $order = new Order($idOrder);
@@ -64,7 +65,7 @@ class AdminDPDBalticsAjaxShipmentsController extends AbstractAdminController
                 $this->changeReceiverAddressBlock($receiverAddressData, $idOrder);
                 break;
             case 'updateAddressBlock':
-                $idAddressDelivery = (int)Tools::getValue('id_address_delivery');
+                $idAddressDelivery = (int) Tools::getValue('id_address_delivery');
                 $this->updateAddressBlock($order, $idAddressDelivery);
                 break;
             case 'print':
@@ -72,11 +73,16 @@ class AdminDPDBalticsAjaxShipmentsController extends AbstractAdminController
                 if ($isAbove177) {
                     $this->returnResponse(['status' => true]);
                 }
-                $shipmentId = (int)Tools::getValue('shipment_id');
+                $shipmentId = (int) Tools::getValue('shipment_id');
                 $labelFormat = Tools::getValue('labelFormat');
                 $labelPosition = Tools::getValue('labelPosition');
                 $this->returnResponse($this->printLabel($shipmentId, $labelFormat, $labelPosition));
                 break;
+            case 'print-return':
+                $response['status'] = false;
+                $shipmentId = (int) Tools::getValue('shipment_id');
+                $this->printReturnLabel($shipmentId);
+                die();
             case 'save':
             case 'save_and_print':
                 $shipmentData = $formDataConverter->convertShipmentFormDataToShipmentObj($data);
@@ -92,7 +98,7 @@ class AdminDPDBalticsAjaxShipmentsController extends AbstractAdminController
                     $product = new DPDProduct($productId);
                     $carrier = Carrier::getCarrierByReference($product->id_reference);
                 } else {
-                    $carrierId = (int)Tools::getValue('carrier_id');
+                    $carrierId = (int) Tools::getValue('carrier_id');
                     $carrier = new Carrier($carrierId);
                 }
                 /** @var PudoService $pudoService */
@@ -124,7 +130,7 @@ class AdminDPDBalticsAjaxShipmentsController extends AbstractAdminController
                 $this->returnResponse(
                     [
                         'status' => true,
-                        'isPudo' => (bool)$dpdProduct->is_pudo
+                        'isPudo' => (bool) $dpdProduct->is_pudo
                     ]
                 );
                 break;
@@ -221,7 +227,7 @@ class AdminDPDBalticsAjaxShipmentsController extends AbstractAdminController
 
     private function changeReceiverAddressBlock($receiverAddressData, $orderId)
     {
-        if (preg_match('#[^0-9]#',$receiverAddressData->phone)) {
+        if (preg_match('#[^0-9]#', $receiverAddressData->phone)) {
             $response['message'] = $this->module->l('Invalid phone number.');
             $this->returnResponse($response);
         }
@@ -281,9 +287,45 @@ class AdminDPDBalticsAjaxShipmentsController extends AbstractAdminController
 
     private function printLabel($shipmentId, $labelFormat, $labelPosition)
     {
+        /** @var LabelPrintingService $labelPrintingService */
         $labelPrintingService = $this->module->getModuleContainer('invertus.dpdbaltics.service.label.label_printing_service');
 
         return $labelPrintingService->setLabelOptions($shipmentId, $labelFormat, $labelPosition);
+    }
+
+    private function printReturnLabel($shipmentId)
+    {
+        /** @var LabelApiService $labelApiService */
+        $labelApiService = $this->module->getModuleContainer('invertus.dpdbaltics.service.api.label_api_service');
+        $returnTemplateId = Tools::getValue('return_template_id') ?: self::DEFAULT_LABEL_TEMPLATE_ID;
+
+        $dpdShipment = new DPDShipment($shipmentId);
+        if ($dpdShipment->return_pl_number) {
+            try {
+                $labelApiService->printLabel($dpdShipment->return_pl_number, false, false, true);
+                exit();
+            } catch (Exception $e) {
+                $response['status'] = false;
+                $this->returnResponse($response);
+            }
+        }
+        try {
+            /** @var ShipmentService $shipmentService */
+            $shipmentService = $this->module->getModuleContainer('invertus.dpdbaltics.service.shipment_service');
+            $dpdShipment = $shipmentService->createReturnServiceShipment($returnTemplateId, $dpdShipment->id_order);
+            $labelApiService->printLabel($dpdShipment->return_pl_number, false, false, true);
+            exit();
+        } catch (DPDBalticsAPIException $e) {
+            /** @var ExceptionService $exceptionService */
+            $exceptionService = $this->module->getModuleContainer('invertus.dpdbaltics.service.exception.exception_service');
+            $errorMessage = $exceptionService->getErrorMessageForException(
+                $e,
+                $exceptionService->getAPIErrorMessages()
+            );
+            $this->context->cookie->dpd_error = json_encode($errorMessage);
+            //todo: it should display error somehow but now it just shows error in empty page.
+            exit($errorMessage);
+        }
     }
 
     private function setLabelOptions(ShipmentData $shipmentData, $shipmentId, $orderId)