+++ controllers/admin/AdminDPDBalticsShipmentSettingsController.php
@@ -1,4 +1,23 @@
 <?php
+/**
+ * Copyright since 2007 PrestaShop SA and Contributors
+ * PrestaShop is an International Registered Trademark & Property of PrestaShop SA
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Academic Free License version 3.0
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * https://opensource.org/licenses/AFL-3.0
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@prestashop.com so we can send you a copy immediately.
+ *
+ * @author    PrestaShop SA and Contributors <contact@prestashop.com>
+ * @copyright Since 2007 PrestaShop SA and Contributors
+ * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
+ */
+
 
 use Invertus\dpdBaltics\Config\Config;
 use Invertus\dpdBaltics\Controller\AbstractAdminController;
@@ -185,6 +204,13 @@ class AdminDPDBalticsShipmentSettingsController extends AbstractAdminController
                         'title' => $this->l('Pickup map'),
                         'validation' => 'isBool',
                         'cast' => 'intval',
+                        'type' => 'bool',
+                    ],
+                    Config::AUTOMATED_PARCEL_RETURN => [
+                        'title' => $this->trans('Automatically generate return label'),
+                        'validation' => 'isBool',
+                        'cast' => 'intval',
+                        'hint' => $this->trans('Enabling automated parcel returns will generate a return label for every DPD order', [], 'Admin.Shipment'),
                         'type' => 'bool',
                     ],
                     Config::GOOGLE_API_KEY => [