+++ src/Install/Installer.php
@@ -1,4 +1,23 @@
 <?php
+/**
+ * Copyright since 2007 PrestaShop SA and Contributors
+ * PrestaShop is an International Registered Trademark & Property of PrestaShop SA
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Academic Free License version 3.0
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * https://opensource.org/licenses/AFL-3.0
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@prestashop.com so we can send you a copy immediately.
+ *
+ * @author    PrestaShop SA and Contributors <contact@prestashop.com>
+ * @copyright Since 2007 PrestaShop SA and Contributors
+ * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
+ */
+
 
 namespace Invertus\dpdBaltics\Install;
 
@@ -303,7 +322,8 @@ class Installer
         }
         /** @var TabsCollection $tabCollection */
         $tabCollection = $this->tabFactory->getTabsCollection();
-        $tabsInitializer = new TabsInitializer(_PS_VERSION_, $tabCollection);
+        $moduleName = $this->module->name;
+        $tabsInitializer = new TabsInitializer(_PS_VERSION_, $tabCollection, $moduleName);
 
         return $tabsInitializer->initializeTabsByPsVersion();
     }
@@ -414,7 +434,6 @@ class Installer
             'displayCarrierExtraContent',
             'actionFrontControllerSetMedia',
             'displayProductPriceBlock',
-            'actionObjectOrderUpdateAfter',
             'actionValidateOrder',
             'displayAdminOrder',
             'displayAdminOrderTabContent',
@@ -423,14 +442,10 @@ class Installer
             'displayOrderDetail',
             'displayBackOfficeTop',
             'displayOrderDetail',
-            'actionDispatcher',
             'actionAdminOrdersListingFieldsModifier',
             'displayAdminListBefore',
-            'actionCarrierProcess',
             'displayAdminOrderTabContent',
             'actionOrderGridDefinitionModifier',
-            'actionObjectCarrierUpdateAfter',
-            'actionObjectCarrierUpdateBefore',
         ];
     }
 }