+++ src/Validate/Phone/PhoneNumberValidator.php
@@ -1,4 +1,23 @@
 <?php
+/**
+ * Copyright since 2007 PrestaShop SA and Contributors
+ * PrestaShop is an International Registered Trademark & Property of PrestaShop SA
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Academic Free License version 3.0
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * https://opensource.org/licenses/AFL-3.0
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@prestashop.com so we can send you a copy immediately.
+ *
+ * @author    PrestaShop SA and Contributors <contact@prestashop.com>
+ * @copyright Since 2007 PrestaShop SA and Contributors
+ * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
+ */
+
 
 namespace Invertus\dpdBaltics\Validate\Phone;
 
@@ -24,31 +43,39 @@ class PhoneNumberValidator
 
     public function isPhoneValid($prefix, $phone)
     {
-        if (empty($prefix)) {
+        if (empty($prefix) && empty($phone)) {
             throw new DpdCarrierException(
-                'Phone number prefix is empty',
-                Config::ERROR_BAD_PHONE_NUMBER_PREFIX
+                'Phone number details are empty',
+                Config::ERROR_EMPTY_PHONE_DETAILS
             );
-        }
-        if (empty($phone)) {
-            throw new DpdCarrierException(
-             'Phone number is empty',
-                Config::ERROR_PHONE_EMPTY
-            );
-        }
-       if (!is_numeric($phone)) {
-           throw new DpdCarrierException(
-              'Phone number contains invalid characters',
-               Config::ERROR_PHONE_HAS_INVALID_CHARACTERS
-           );
-       }
+        } else {
 
-       if (strlen($phone) < 8 || strlen($phone) > 13) {
-           throw new DpdCarrierException(
-              'Phone number length is invalid',
-               Config::ERROR_PHONE_HAS_INVALID_LENGTH
-           );
-       }
+            if (empty($prefix)) {
+                throw new DpdCarrierException(
+                    'Phone number prefix is empty',
+                    Config::ERROR_BAD_PHONE_NUMBER_PREFIX
+                );
+            }
+            if (empty($phone)) {
+                throw new DpdCarrierException(
+                    'Phone number is empty',
+                    Config::ERROR_PHONE_EMPTY
+                );
+            }
+            if (!is_numeric($phone)) {
+                throw new DpdCarrierException(
+                    'Phone number contains invalid characters',
+                    Config::ERROR_PHONE_HAS_INVALID_CHARACTERS
+                );
+            }
+
+            if (strlen($phone) < 6 || strlen($phone) > 13) {
+                throw new DpdCarrierException(
+                    'Phone number length is invalid',
+                    Config::ERROR_PHONE_HAS_INVALID_LENGTH
+                );
+            }
+        }
 
         return true;
     }