+++ views/js/front/order-opc.js
@@ -1,3 +1,21 @@
+/**
+ * Copyright since 2007 PrestaShop SA and Contributors
+ * PrestaShop is an International Registered Trademark & Property of PrestaShop SA
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Academic Free License version 3.0
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * https://opensource.org/licenses/AFL-3.0
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@prestashop.com so we can send you a copy immediately.
+ *
+ * @author    PrestaShop SA and Contributors <contact@prestashop.com>
+ * @copyright Since 2007 PrestaShop SA and Contributors
+ * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
+ */
 var isPudoPointSelected = false;
 
 $(document).ready(function (){
@@ -26,7 +44,11 @@ function handlePhoneNumber(selector)
     var phone = selector.find('input[name="dpd-phone"]').val();
     var phoneArea = selector.find('select[name="dpd-phone-area"] option:selected').val();
 
-    saveSelectedPhoneNumber(phone, phoneArea)
+    if (currentController === 'supercheckout') {
+        saveSelectedPhoneNumberSuperCheckout(phone, phoneArea)
+    } else {
+        saveSelectedPhoneNumber(phone, phoneArea)
+    }
 
     return true;
 }