+++ CHANGELOG.md
@@ -114,3 +114,16 @@
 - API endpoint changed in dpdbalticsapi, logic improved, no to send unnecessary params
 - Remark message upgrade added, allow special chars, trim content to prevent error.
 
+## [3.2.10]
+ - Validation message styles unified with super checkout module
+ - Validation message handling added(remove change and display correct messages)
+ - Logic to execute superchekout handling added only when correct carrier selected
+ - Parcel return label logic added to Admin Order View
+ - Automated parcel return settings added to Shipment settings
+ - On Automated parcel setting return Shipment and Return labels are printed together
+
+## [3.2.11]
+- Sustainable logo added to DPD carrier text for older versions of theme
+
+## [3.2.12]
+- Print return label on initial order creation in bulk action and save&print bug fix